Refactor to move processing out of matfact
