+++ matfact/pyproject.toml
@@ -18,6 +18,7 @@ mlflow = "^1.28.0"
 tqdm = "^4.64.1"
 scikit-optimize = "^0.9.0"
 pydantic = {extras = ["dotenv"], version = "^1.10.4"}
+processing = {path = "../processing", develop = true}
 
 [tool.poetry.group.dev.dependencies]
 ipykernel = "^6.15.2"
@@ -43,6 +44,7 @@ plugins = "numpy.typing.mypy_plugin"
 [[tool.mypy.overrides]]
 module = [
     "matfact.*",
+    "processing.*",
     "matplotlib.*",
     #
     "scipy.*",
