+++ processing/processing/pipelines.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 from sklearn.pipeline import Pipeline
 
-from matfact.processing.transformers import (
+from processing.transformers import (
     AgeAdder,
     AgeBinAssigner,
     BirthdateAdder,
@@ -12,7 +12,6 @@ from matfact.processing.transformers import (
     RiskAdder,
     RowAssigner,
 )
-
 
 def matfact_pipeline(
     *,
